feat: stub out components for crowdsourcing form
Rename `AddTapModal` to `AddResourceModal`

Move it into its own directory so we can keep all related components
and files in one place.

Added `formStep` field to `AddResourceModal` state so we can
dynamically show different modal components based on the state of
`formStep`

`formStep` is a string that indicates the "step" in the form
that the user is currently at. We can name the steps in a descriptive
manner so they describe what the user is doing, and the associated
modal UI they should display.

E.g. when `formStep` is set to `"chooseResource"`, we can show the
components which is the set of buttons that lets the user choose
what resource they would like to add.

Additionally, we create a bunch of additional components
within the directory to help us break out the different views
we have for adding different resource types