fix: move submit button so it's linked with form component
Buttons for resource modals lived in `Modal.Footer` component outside of
react-bootstrap `Form` component, resulting in the submit button not
triggering `onSubmit` attribute in `Form` component.

To fix, we move the `submit` button into the form component.

Some extra styles have been added to accomodate these changes, but
as a result buttons do not look close to the mockup