Fix map panning issues related resource info modal (#539)
* Set user location after location fetch

* Newer action syntax for user location, map center

* Correct info window class on open

* Remove old, commented out action syntax

* Fixes a map panning issue from issue #527

This fixes the issue where opening a site, then clicking "Near Me", then
closing the info modal was panning back to the prior site, whereas it
should have stayed at the "Near Me" site. This is "issue C" as discussed
in the comments of GitHub issue #527.

* Rename redux state mapCenter to lastResourcePan

The name change reflects the fact that currently as a user may have
panned around the map by clicking and dragging the map, the `mapCenter`
state was not updated. Instead, the state is updated when e.g. a pin
is clicked, or the "Near Me" button is clicked.

* Remove debugging statement

* Restore original "Near Me" behavior after fix

* Add some documentation comments

* Simplify action - setSelectedPlace

Removed unnecessary if-else condition from the reducer, and
transitioned to the `createAction` syntax.

* Remove temp port change for dev

* Small bit of cleanup ahead of merging

---------

Co-authored-by: Gabriel Cardona <2278918+gcardonag@users.noreply.github.com>